trying to fix people page bug with loading
